bugfix for rev. 3207 : pimcore memcache and mysqltable cache backend bugfixes ;-)
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@3209 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/lib/Pimcore/Cache/Backend/Memcached.php
@@ -232,7 +232,9 @@ class Pimcore_Cache_Backend_Memcached extends Zend_Cache_Backend_Memcached {
                     $quotedIds[] = $this->getDb()->quote($item);
                 }
                 //$this->getDb()->delete("cache_tags", "tag = '".$tag."'");
-                $this->getDb()->delete("cache_tags", "id IN (" . implode(",", $quotedIds) . ")");
+                if(count($quotedIds) > 0) {
+                    $this->getDb()->delete("cache_tags", "id IN (" . implode(",", $quotedIds) . ")");
+                }
             }
         }
         if ($mode == Zend_Cache::CLEANING_MODE_NOT_MATCHING_TAG) {
